Rename ConveratinAnalysisProjectsClient to ConversationAuthoringClient
