+++ sdk/cognitivelanguage/Azure.AI.Language.Conversations/src/autorest.md
@@ -12,6 +12,7 @@ batch:
   clear-output-folder: true
 
 - input-file: https://raw.githubusercontent.com/Azure/azure-rest-api-specs/725f4ca360426a32d20e81eb945065e62c285d6a/specification/cognitiveservices/data-plane/Language/stable/2022-05-01/analyzeconversations-authoring.json
+  namespace: Azure.AI.Language.Conversations.Authoring
 
 data-plane: true
 model-namespace: false
