+++ sdk/cognitivelanguage/Azure.AI.Language.Conversations/tests/ConversationAuthoringClientTests.cs
@@ -2,27 +2,28 @@
 // Licensed under the MIT License.
 
 using System;
+using Azure.AI.Language.Conversations.Authoring;
 using NUnit.Framework;
 
 namespace Azure.AI.Language.Conversations.Tests
 {
-    public class ConversationAnalysisClientTests
+    public class ConversationAuthoringClientTests
     {
         [Test]
-        public void ConversationAnalysisClientEndpointNull()
+        public void ConversationAuthoringClientEndpointNull()
         {
             ArgumentNullException ex = Assert.Throws<ArgumentNullException>(
-                () => new ConversationAnalysisClient(null, null));
+                () => new ConversationAuthoringClient(null, null));
             Assert.AreEqual("endpoint", ex.ParamName);
         }
 
         [Test]
-        public void ConversationAnalysisClientCredentialNull()
+        public void ConversationAuthoringClientCredentialNull()
         {
             Uri endpoint = new("https://test.cognitive.microsoft.com", UriKind.Absolute);
 
             ArgumentNullException ex = Assert.Throws<ArgumentNullException>(
-                () => new ConversationAnalysisClient(endpoint, null));
+                () => new ConversationAuthoringClient(endpoint, null));
             Assert.AreEqual("credential", ex.ParamName);
         }
     }
