+++ sdk/cognitivelanguage/Azure.AI.Language.Conversations/tests/SessionRecords/ConversationAuthoringClientLiveTests/GetSupportedMultilingualPrebuiltEntities.json
