+++ sdk/cognitivelanguage/Azure.AI.Language.Conversations/tests/SessionRecords/ConversationAuthoringClientLiveTests/GetSupportedMultilingualPrebuiltEntitiesAsync.json
